+++ tests/fixtures/split-php/namespace_use_group/aliased.out.ts
@@ -1,0 +1,4 @@
+<?php use App\{
+  Foo as F,
+  Bar
+};