+++ tests/fixtures/split-recursive-php/call-with-array.out.ts
@@ -1,0 +1,10 @@
+<?php route(
+  [
+    'GET',
+    'POST',
+  ],
+  handle(
+    $a,
+    $b,
+  ),
+);